requirements.txt
prometheus_client
dnspython
requests
elasticsearch~=5.0
